fix ai21 deprecated models
