+++ tests/local_testing/test_completion.py
@@ -3940,24 +3940,6 @@ def test_mistral_anyscale_stream():
 #         pytest.fail(f"Error occurred: {e}")
 
 
-#### Test A121 ###################
-@pytest.mark.skip(reason="Local test")
-def test_completion_ai21():
-    print("running ai21 j2light test")
-    litellm.set_verbose = True
-    model_name = "j2-light"
-    try:
-        response = completion(
-            model=model_name, messages=messages, max_tokens=100, temperature=0.8
-        )
-        # Add any assertions here to check the response
-        print(response)
-    except Exception as e:
-        pytest.fail(f"Error occurred: {e}")
-
-
-# test_completion_ai21()
-# test_completion_ai21()
 ## test deep infra
 @pytest.mark.parametrize("drop_params", [True, False])
 def test_completion_deep_infra(drop_params):
